Update Base.css
small fix for border-box and padding-box 

Won't have much impact in production yet because this is still experimental technology. 

I noticed that this was the upstream project that we inherited for DownmarkerWPF so I thought I should submit a pull request here as well. Approved and merged at https://github.com/Code52/DownmarkerWPF/pull/417 


Thanks! 

https://developer.mozilla.org/en-US/docs/Web/CSS/box-sizing